開発時用のファイルを削除し整理した
