--- MotionDataVerification/Assets/_PremaidAI/HumanoidTracer/Scenes/HumanoidMimicryTest.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 23f23a9d453ec804480df95b85de5f55
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
