--- MotionDataVerification/Assets/_PremaidAI/HumanoidTracer/Scenes/HumanoidTracerMimic.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 38065eee418d36d40a3dd3d02789b22b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
