--- MotionDataVerification/Assets/_PremaidAI/HumanoidTracer/Scripts/HumanoidTracerManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 282f0475dc0e0ca49a23e4a1b9405216
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
